PLAT-10459: Simplify examples
Turn the examples in simple scripts under one folder, debug logging in
the console enabled by default.
